Add new config handling feature

+++ main.ps1
@@ -11,48 +11,28 @@ catch {
 }
 }
 
-# Override for the confirm function
-$confirmOverride = $false
-
-function Confirmation{
-	param ([string]$text)
-
-	if ($confirmOverride) {
-		return $true
- 	}
-
-	$option = (Read-Host $text "[y/n]").ToLower()
-	if (!(($option -eq "y") -Or ($option -eq "yes")) -And !(($option -eq "n") -Or ($option -eq "no"))) {
-		return Confirmation $text
-	}
-
-    	return ($option -eq "y") -Or ($option -eq "yes")
-}
-
-# Option to accept all the options
-if (Confirmation "Apply all options") {
-	$confirmOverride = $true
-}
-
 # Download user config file
 try {
-    $configFileUrl = "https://github.com/likes-gay/win-config/blob/main/configs/{0}.xml"-f $Env:UserName
-    Invoke-WebRequest $configFileUrl -outfile "config.xml"
+    $configFileUrl = "https://github.com/likes-gay/win-config/blob/main/configs/{0}.json"-f $Env:UserName
+    Invoke-WebRequest $configFileUrl -outfile "config.json"
 
 } catch {
     Write 'No config file detected, please create one in this folder: https://github.com/likes-gay/win-config/blob/main/configs/'
     Exit
 }
 
+# Parse config file
+$configFile = Get-Content .\config.json -Raw | ConvertFrom-Json
+
 # Unpin unused apps from the taskbar
-if (Confirmation "Unpin unused apps") {
+if ($configFile.'Unpin-apps') {
 	UnPin-App "Microsoft Edge"
 	UnPin-App "Microsoft Store"
 	UnPin-App "Mail"
 }
 
 # Turns on dark mode for apps and system
-if (Confirmation "Turn on dark mode for apps and system") {
+if ($configFile.'Dark-mode') {
 	$themesPersonalise = "HKCU:\SOFTWARE\Microsoft\Windows\CurrentVersion\Themes\Personalize"
 	Set-ItemProperty -Path $themesPersonalise -Name "AppsUseLightTheme" -Value 0 -Type Dword
 	Set-ItemProperty -Path $themesPersonalise -Name "SystemUsesLightTheme" -Value 0 -Type Dword
@@ -61,42 +41,43 @@ if (Confirmation "Turn on dark mode for apps and system") {
 $explorer = "HKCU:\SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\Advanced"
 
 # Remove task view
-if (Confirmation "Remove task view") {
+if ($configFile.'Remove-task-view') {
 	Set-ItemProperty -Path $explorer -Name "ShowTaskViewButton" -Value 0
 }
 
 # Turn on file extensions in File Explorer
-if (Confirmation "Turn on file extensions in File Explorer") {
+if ($configFile.'File-extentions') {
 	Set-ItemProperty -Path $explorer -Name "HideFileExt" -Value 0
 }
 
 # Hide desktop icons
-if (Confirmation "Hide desktop icons") {
+if ($configFile.'Remove-desktop-icons') {
 	Set-ItemProperty -Path $explorer -Name "HideIcons" -Value 1
 }
 
 # Enable seconds in clock
-if (Confirmation "Enable seconds on clock") {
+if ($configFile.'Seconds-in-clock') {
 	Set-ItemProperty -Path $explorer -Name "ShowSecondsInSystemClock" -Value 1 -Force
 }
 
 # Enable 12 hour time in clock
-if (Confirmation "Enable 12 hour time in clock") {
+if ($configFile.'12-hr-clock') {
 	Set-ItemProperty -Path $explorer -Name "UseWin32TrayClockExperience" -Value 0 -Force
 }
 
 # Enable the clipboard history
-if (Confirmation "Enable clipboard history") {
+if ($configFile.'clipboard-history') {
 	Set-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Clipboard" -Name "EnableClipboardHistory" -Value 1
 }
 
 # Set print screen to open snipping tool
-if (Confirmation "Rebind print screen to open snipping tool") {
+if ($configFile.'Print-scrn-snipping-tool') {
 	Set-ItemProperty -Path "HKCU:\Control Panel\Keyboard" -Name "PrintScreenKeyForSnippingEnabled" -Value 1 -Type Dword
 }
 
 # Set scroll lines to 7
-if (Confirmation "Set scroll lines to 7") {
+if ($configFile.'Set-scroll-lines') {
+    $scrollSpeed = $configFile.'Set-scroll-lines'
 	Add-Type -TypeDefinition @"
 using System;
 using System.Runtime.InteropServices;
@@ -110,67 +91,75 @@ public class WinAPI {
 }
 "@
 
-	$scrollSpeed = 7
 	Set-ItemProperty -Path "HKCU:\Control Panel\Desktop" -Name "WheelScrollLines" -Value $scrollSpeed
 	[WinAPI]::SystemParametersInfo(0x0069, $scrollSpeed, 0, 2)
 	[WinAPI]::SendMessageTimeout(0xffff, 0x1a, [IntPtr]::Zero, "Environment", 2, 5000, [IntPtr]::Zero)
 }
 
 # Turn on "Live Caption" in Google Chrome
-if (Confirmation "Turn on 'Live Caption' in Google Chrome") {
+if ($configFile.'Enable-live-caption-chrome') {
 	$originalFile = "$env:LocalAppData\Google\Chrome\User Data\Default\Preferences"
 	$content = Get-Content -Path $originalFile | ConvertFrom-Json
 	$content.accessibility.captions.live_caption_enabled = "true"
 	$content | ConvertTo-Json -Compress | Set-Content -Path $originalFile
+    Set-Content -Path $originalFile -Value $content
 }
 
 # Set default browser to Chrome
-Invoke-WebRequest  "https://raw.githubusercontent.com/likes-gay/win-config/main/default_browser.vbs" -OutFile .\default_browser.vbs
-Invoke-Expression "Cscript.exe .\default_browser.vbs //nologo"
-Remove-Item -Path ".\default_browser.vbs"
+if ($configFile.'Default-browser-chrome') {
+    Invoke-WebRequest  "https://raw.githubusercontent.com/likes-gay/win-config/main/default_browser.vbs" -OutFile .\default_browser.vbs
+    Invoke-Expression "Cscript.exe .\default_browser.vbs //nologo"
+    Remove-Item -Path ".\default_browser.vbs"
 
-# Setup edge redirect - https://github.com/rcmaehl/MSEdgeRedirect/wiki/Deploying-MSEdgeRedirect
-if (Confirmation "Install and configure MSEdgeRedirect") {
-	Invoke-WebRequest "https://github.com/rcmaehl/MSEdgeRedirect/releases/latest/download/MSEdgeRedirect.exe" -OutFile .\MSEdgeRedirect.exe
-	Invoke-WebRequest "https://raw.githubusercontent.com/likes-gay/win-config/main/edge_redirect.ini" -OutFile .\edge_redirect.ini
-	Start-Process "MSEdgeRedirect.exe" -ArgumentList "/silentinstall",".\edge_redirect.ini" -PassThru
-	Remove-Item -Path ".\edge_redirect.ini"
-	Remove-Item -Path ".\MSEdgeRedirect.exe"
+    # Setup edge redirect - https://github.com/rcmaehl/MSEdgeRedirect/wiki/Deploying-MSEdgeRedirect
+    if ($configFile.'Setup-edge-redirect') {
+	    Invoke-WebRequest "https://github.com/rcmaehl/MSEdgeRedirect/releases/latest/download/MSEdgeRedirect.exe" -OutFile .\MSEdgeRedirect.exe
+	    Invoke-WebRequest "https://raw.githubusercontent.com/likes-gay/win-config/main/edge_redirect.ini" -OutFile .\edge_redirect.ini
+	    Start-Process "MSEdgeRedirect.exe" -ArgumentList "/silentinstall",".\edge_redirect.ini" -PassThru
+	    Remove-Item -Path ".\edge_redirect.ini"
+	    Remove-Item -Path ".\MSEdgeRedirect.exe"
+    }
 }
 
-try {
-	Stop-Process -Name msedge -Force
-} catch {
-	Write-Output "Microsoft Edge is already shut"
+
+if ($configFile.'Close-edge'){
+    try {
+	    Stop-Process -Name msedge -Force
+    } catch {
+	    Write-Output "Microsoft Edge is already shut"
+    }
 }
 
-Set-Content -Path $originalFile -Value $content
 
 Stop-Process -processName: Explorer # Restart explorer to apply changes that require it
 
-# Open useful tabs
-Start-Process "chrome.exe" "https://www.bbc.co.uk/news"
-Start-Process "chrome.exe" "https://github.com/login"
-Start-Process "chrome.exe" "https://office.com"
-Start-Process "chrome.exe" "https://teams.microsoft.com/v2" -Wait -PassThru
-
-# Easter egg ;)
-$images = (Invoke-WebRequest "https://raw.githubusercontent.com/likes-gay/win-config/main/photos.txt").Content.Split([Environment]::NewLine)
-
-# Create folder to store downloaded images in to prevent clutter.
-$downloadPath = $env:USERPROFILE + "\Downloads\likes-gay-images"
-If (!(test-path $downloadPath)) {
-	New-Item -ItemType Directory -Path $downloadPath
+if ($configFile.'Open-tabs') {
+    # Open useful tabs
+    Start-Process "chrome.exe" "https://www.bbc.co.uk/news"
+    Start-Process "chrome.exe" "https://github.com/login"
+    Start-Process "chrome.exe" "https://office.com"
+    Start-Process "chrome.exe" "https://teams.microsoft.com/v2" -Wait -PassThru
 }
 
-foreach ($i in $images) {
-	# Get the name of the image from the URL
-	# Windows will not open images in the photo viewer unless they have a file extension.
-	$imageName = $i.split("/")[$i.split("/").Count - 1]
+if ($configFile.'Funny-joe-biden'){
+    # Easter egg ;)
+    $images = (Invoke-WebRequest "https://raw.githubusercontent.com/likes-gay/win-config/main/photos.txt").Content.Split([Environment]::NewLine)
 
-	# Download and open the image
-	Invoke-WebRequest -Uri $i -OutFile $downloadPath\$imageName
-	Start-Process $downloadPath\$imageName
+
+    # Create folder to store downloaded images in to prevent clutter.
+    $downloadPath = $env:USERPROFILE + "\Downloads\likes-gay-images"
+    If (!(test-path $downloadPath)) {
+	    New-Item -ItemType Directory -Path $downloadPath
+    }
+
+    foreach ($i in $images) {
+	    # Get the name of the image from the URL
+	    # Windows will not open images in the photo viewer unless they have a file extension.
+	    $imageName = $i.split("/")[$i.split("/").Count - 1]
+
+	    # Download and open the image
+	    Invoke-WebRequest -Uri $i -OutFile $downloadPath\$imageName
+	    Start-Process $downloadPath\$imageName
+    }
 }
-
 exit
